+++ src/main/kotlin/kyta/composter/protocol/packet/play/ClientboundRemoveEntityPacket.kt
@@ -3,7 +3,7 @@ package kyta.composter.protocol.packet.play
 import kyta.composter.protocol.Packet
 import kyta.composter.protocol.PacketSerializer
 import kyta.composter.protocol.WriteBuffer
-import xyz.nkomarn.composter.entity.Entity
+import kyta.composter.world.entity.Entity
 
 data class ClientboundRemoveEntityPacket(val id: Int) : Packet {
     constructor(entity: Entity) : this(entity.id)
