+++ src/main/kotlin/kyta/composter/protocol/packet/play/EntityPositionPackets.kt
@@ -1,7 +1,7 @@
 package kyta.composter.protocol.packet.play
 
-import kyta.composter.asAbsoluteInt
-import kyta.composter.asRotation
+import kyta.composter.protocol.asAbsoluteInt
+import kyta.composter.protocol.asRotation
 import kyta.composter.protocol.Packet
 import kyta.composter.protocol.PacketSerializer
 import kyta.composter.protocol.WriteBuffer
