+++ src/main/kotlin/kyta/composter/server/MinecraftServer.kt
@@ -3,16 +3,15 @@ package kyta.composter.server
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Runnable
 import kotlinx.coroutines.asCoroutineDispatcher
-import kyta.composter.Tickable
 import kyta.composter.network.NetworkController
 import net.kyori.adventure.text.flattener.ComponentFlattener
 import net.kyori.adventure.text.serializer.plain.PlainTextComponentSerializer
 import org.slf4j.LoggerFactory
 import xyz.nkomarn.composter.Composter
-import xyz.nkomarn.composter.server.WorldManager
 import java.nio.file.Paths
 import java.util.concurrent.ConcurrentLinkedDeque
 import java.util.concurrent.Executor
+import kyta.composter.server.world.WorldManager
 
 class MinecraftServer(val composter: Composter) : Tickable, CoroutineScope {
     override val coroutineContext = Executor { executeServerTask(it) }.asCoroutineDispatcher()
