+++ src/main/kotlin/kyta/composter/server/Tickable.kt
@@ -1,4 +1,4 @@
-package kyta.composter
+package kyta.composter.server
 
 interface Tickable {
     fun tick(currentTick: Long)
