+++ src/main/kotlin/kyta/composter/server/helpers.kt
@@ -1,15 +1,7 @@
-package kyta.composter
+package kyta.composter.server
 
 import kotlinx.coroutines.CoroutineScope
 
 suspend fun <T> withContext(scope: CoroutineScope, block: suspend CoroutineScope.() -> T): T {
     return kotlinx.coroutines.withContext(scope.coroutineContext, block)
-}
-
-fun Double.asAbsoluteInt(): Int {
-    return (this * 32.0).toInt()
-}
-
-fun Float.asRotation(): Int {
-    return (this * 256F / 360F).toInt()
 }
