+++ src/main/kotlin/kyta/composter/world/block/Block.kt
@@ -1,6 +1,6 @@
 package kyta.composter.world.block
 
-data class Block(val id: Int)
+data class Block(val networkId: Int)
 
 data class BlockState(
     val block: Block,
