+++ src/main/kotlin/kyta/composter/world/entity/Player.kt
@@ -1,10 +1,8 @@
-package xyz.nkomarn.composter.entity
+package kyta.composter.world.entity
 
 import java.util.function.Consumer
 import kyta.composter.container.BasicContainer
 import kyta.composter.container.menu.MenuSynchronizer
-import kyta.composter.entity.EntityType
-import kyta.composter.entity.ItemEntity
 import kyta.composter.item.Item
 import kyta.composter.item.ItemStack
 import kyta.composter.item.isEmpty
@@ -21,7 +19,7 @@ import kyta.composter.world.ChunkPos
 import kyta.composter.world.World
 import kyta.composter.world.getCollidingEntities
 import net.kyori.adventure.text.Component
-import xyz.nkomarn.composter.entity.tracker.EntityTracker
+import kyta.composter.server.world.entity.tracker.EntityTracker
 
 class Player(
     world: World,
