--- src/main/resources/composter.yml
@@ -1,18 +1,0 @@
-server:
-  network:
-    host: 0.0.0.0
-    port: 25565
-  slots: 200
-
-messages:
-  already_logged_in: "Already logged in."
-  already_handshook: "Already handshook."
-  unsupported_protocol: "Composter only supports Beta 1.7.3."
-
-world:
-  seed: 1
-
-# Supported for Beta 1.8 - Release 1.6.4 server list pings
-motd:
-  enabled: true
-  message: "Composter - Beta 1.7.3"
